dts: use auart4 also in development mode
duart got no DMA and is named differently in userspace.
Make things more consistent by always using auart4, but in development
mode also using it as the primary console for Linux.
The result should be similar to having duart routed to the external
pins, just that now the serial device is named ttyAPP4 and DMA
operation should work.

Fixes #6
Fixes #9

Signed-off-by: Daniel Golle <daniel@makrotopia.org>